Exercices/gcp_core/gcs/solutions/requirements.txt
tqdm
google-cloud-storage==2.7.0